00_uv_litellm_crewai/03_explore_crewai/pyproject.toml
[project]
name = "explore-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Asharib", email = "asharibali@proton.me" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.95.0",
    "litellm>=1.60.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
list-billionares = "explore_crewai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
